In method 'type_query':
Added a "when" clause for columns of type :decimal so that we don't
try to use the LIKE operator on a numeric type, which gives an error
in PostgreSQL.  This fixes redmine issue #1909.

Made text searches case-insensitive by apply LOWER to each of the
operands.  This fixes part of redmine issue #1906.