In display_csv_file:
Make error a local variable

Close uploaded file after writing it so CSV.open doesn't zero it out.

Factor out code that checks the CSV file is well-formed.

Set errors variable to the exception message, not the object itself so as not to overload the cookie store.

Put errors into the flash and redirect to start page instead of just rendering its template.

In insert_data:

Use transactions for data insertion and catch any exceptions.

Redirect to previous step with exception info in flash if errors.