In display_csv_file:
  Determine if the page was reached via a form submission (on the
  start_upload page) based upon the new hidden variable "new upload"
  instead of looking at whether the "CSV file" param was set since the
  user might have submitted the form without choosing a file; always
  reset the session values of csvpath and mapping if the user submitted
  the upload file form.

  Restore @headers and @data from session[:csvpath] if we reached this
  page not via start_upload's form submission.

Do validation of the CSV file's format--or not--inside the read_data
method based on the value of an optional flag parameter.

Factor out displayed_columns method.

In get_insert_data, return a new variable @mapped_data instead of
altering @data, and change the confirm_data template and the
insert_data action to use it instead.

Various comment addtions and changes.