Send email on new signup RM-1707
Instead of creating an issue in redmine this will send an email to all
users who are of page_access_level==1 when a new user is created and/or
when the user requests page_access_level < 4 or acces_level < 3.

Make sure to update the database before using this, since all people in
the BETY dump are page_access_level == 1.