Bulk Upload:
Amendment to last commit: Use the existence of session[:rounding] as the
indicator of whether the 'choose_global_data_values' form has ever been
submitted, since session[:global_values] may legitimately be nil if all required
data was included in the upload file.

Added two new tests to check that needed global information has been submitted
and that the user is redirected back to the form if not.

Tightened up a test for successful data insertion.