Many expansions and revisions to RDoc comments.
Implemented rounding of covariate values.

Handle SE and n for trait uploads having a single trait variable.

During validation, if citation information is in the file but site or treatment
information is not, ensure there is a common site and/or common treatment name
for all the citations listed in the file.

Whitespace normalization fix for interactively-entered cultivar.

Fix to data_value_error_count to accomodate new non-row-specific errors.

Fix typo in method name.

Correctly handle the case where citation information is in the file but the
treatment name is entered interactively.