add IDs to many-many tables
This ID will be used to find those tables that should be synchronized
and what data was entered at what site.

The id's are set to auto increment and will be defaulted to be in the
range of EBI.

The script fix.many.id.sh can be used to fix the auto increment ID to
be in the right range and will try to set the ID's of those that are
entered at the remote site (this is done by looking at the id's of the
elements referenced and see if any of them fall inside the range of the
site). To run the scrip use the following syntax: `MYSITE=1
./fix.many.id.sh`. The arguments are the same as load.bety.sh in pean.

@mdietze Currently this script will only need to be executed at BU and
should not be run at EBI.