Adding new "id" column as primary key to PftsPriors and PftsSpecies models. This prevents a bug whereby insertion into the corresponding tables fails in Rails 3.2 because Rails tries to explicitly set these columns to null which violates the not-null database constraint.
Changed PftsController#make_clone to use the dup() method rather than the
clone() method to make a copy of a Pft object.  This is the correct method to
use in Rails 3.1 and above.  Also, append '-copy' to the clone's name so that
new uniqueness constraints requiring (name, model_id) to be unique will not be
violated.