A few changes to get some more spec to pass:
app/controllers/api/v1/traits_controller.rb:

Api::V1::TraitsController#create used to set the response content_type to
“application/xml”.  It seemed that this was overridden by the type of the
template rendered but is no longer.  So we should just not set it.

Gemfile, Gemfile.lock:

Had to add the “rails-controller-testing” to support tests which use the
“assigns” method.

app/controllers/bulk_upload_controller.rb:

In the BulkUploadController#display_csv_file, we have to replace the test “if
uploaded_io” with “if !uploaded_io.blank?” because the call "post
'display_csv_file', { 'CSV file' => nil }” in the bulk upload controller spec
now results in the value of the “CSV file” parameter being “”:String instead of
NilClass.