Eliminated the 100-item limit on showing files, runs, formats, models, and pfts associated with inputs, machines, models, modeltypes, and posteriors.
This had been implemented using a call to #all with argument "limit: 100", which
is an obsolete call signature in Rails 5.1.  Moreover, most of the association
counts are far less than 100, and in any case, it seemed a bad idea to limit the
display to 100 items without a way to page through to additional items not
shown.